requirements.txt
bcrypt                 4.0.1
blinker                1.6.2
click                  8.1.6
colorama               0.4.6
config                 0.5.1
Flask                  2.3.2
Flask-Cors             4.0.0
Flask-Login            0.6.2
itsdangerous           2.1.2
Jinja2                 3.1.2
MarkupSafe             2.1.3
mysql-connector-python 8.0.33
protobuf               3.20.3
python-dotenv          1.0.0
requests               2.28.2
Werkzeug               2.3.6